+++ Client/Frontend/Onboarding/Models/OnboardingCardInfoModel.swift
@@ -1,0 +1,15 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+
+import Foundation
+
+struct OnboardingCardInfoModel {
+    let name: String
+    let title: String
+    let body: String
+    let image: String
+    let link: OnboardingLinkInfo?
+    let buttons: [OnboardingButtonInfo]
+    let type: OnboardingType
+}
